add Trash support

+++ src/components/SelectNote/DeleteNoteButton.res
@@ -1,5 +1,5 @@
 @react.component
-let make = (~note: Shape.Note.t, ~afterDelete) => {
+let make = (~note: Shape.Note.t, ~afterDelete, ~trash=false) => {
   let (isOpen, toggleOpen, _) = Hook.useToggle()
 
   let onClick = evt => {
@@ -7,21 +7,26 @@ let make = (~note: Shape.Note.t, ~afterDelete) => {
     toggleOpen()
   }
 
-  let {deleteNote} = Store.Notes.use()
+  let {deleteNote, updateNote} = Store.Notes.use()
 
   let onDelete = _ => {
-    deleteNote(note.id)
+    if note.isDeleted {
+      deleteNote(note.id)
+    } else {
+      updateNote({...note, isDeleted: true})
+    }
     afterDelete()
-    Toast.success("Note deleted successfully")
+    Toast.success(`Note ${trash ? "trash" : "delet"}ed successfully`)
   }
 
+  let className = trash ? "" : "hover:!bg-error/80 hover:text-error-content"
   <React.Fragment>
-    <button onClick className="hover:!bg-error/80 hover:text-error-content">
-      <Icon.trash className="text-xl" />
-      {"Delete"->React.string}
+    <button onClick className>
+      {trash ? <Icon.trashSimple className="text-xl" /> : <Icon.trash className="text-xl" />}
+      {(trash ? "Trash" : "Delete")->React.string}
     </button>
     {isOpen
-      ? <Modal title="Delete note" onClose=toggleOpen>
+      ? <Modal title={trash ? "Trash note" : "Delete note"} onClose=toggleOpen>
           <div className="flex flex-col gap-4">
             <h2 className="card-title resp-title"> {note.title->React.string} </h2>
             <pre className="line-clamp-4 text-wrap grow -mt-2">
@@ -33,7 +38,7 @@ let make = (~note: Shape.Note.t, ~afterDelete) => {
           </div>
           <div className="flex flex-row gap-4 mt-4 justify-end">
             <button onClick=onDelete className="btn resp-btn btn-error">
-              {React.string("Yes, Delete")}
+              {(trash ? "Yes, Trash" : "Yes, Delete")->React.string}
             </button>
             <button onClick className="btn resp-btn"> {React.string("Cancel")} </button>
           </div>
